chore: complted base of card
