+++ client/package.json
@@ -13,6 +13,8 @@
         "prop-types": "^15.7.2",
         "react": "^16.13.0",
         "react-async": "^10.0.0",
+        "react-circular-progressbar": "^2.0.3",
+        "react-date-picker": "^8.0.1",
         "react-dom": "^16.13.0",
         "react-router-dom": "^5.1.2",
         "react-scripts": "3.4.0",
