+++ client/src/components/GoalsList/GoalsList.scss
@@ -1,0 +1,8 @@
+.grid {
+    display: grid;
+    grid-template-columns: repeat(auto-fit, minmax(340px, 1fr));
+    grid-template-rows: auto;
+    grid-gap: 2rem;
+    justify-items: center;
+    padding: 1.5rem 0.7rem;
+}
