+++ client/src/components/common/Button/Button.js
@@ -1,7 +1,25 @@
 import React from 'react';
 
-function Button({ children, type, className = '', onClick, loading = false }) {
+function Button({
+    children,
+    type,
+    className = '',
+    onClick,
+    loading = false,
+    icon
+}) {
     const btnClass = 'button ' + className + (loading ? ' is-loading ' : '');
+
+    if (icon) {
+        return (
+            <button onClick={onClick} className={btnClass}>
+                <span className="material-icons icon is-small has-margin-right-15 mdi mdi-dark">
+                    {icon}
+                </span>
+                <span className="is-size-6">{children}</span>
+            </button>
+        );
+    }
 
     return (
         <button
