+++ client/src/components/common/Card/CardFooter/CardFooter.js
@@ -1,0 +1,8 @@
+import React from 'react';
+
+function CardFooter({ className = '', children }) {
+    const classNames = 'card-footer is-flex' + className;
+    return <footer className={classNames}>{children}</footer>;
+}
+
+export default CardFooter;
