+++ client/src/components/common/Card/index.js
@@ -1,0 +1,6 @@
+import Card from './Card';
+import CardContent from './CardContent/CardContent';
+import CardFooter from './CardFooter/CardFooter';
+import CardHeader from './CardHeader/CardHeader';
+
+export { Card, CardContent, CardFooter, CardHeader };
