+++ client/src/components/common/CreateGoalForm/CreateGoalForm.scss
@@ -22,6 +22,15 @@ $goal-colors: 'BerryRed', 'Red', 'Orange', 'Yellow', 'OliveGreen', 'LimeGreen',
         border-radius: 50%;
         @include color-modifiers();
     }
+    .react-date-picker__wrapper {
+        box-shadow: inset 0 0.0625em 0.125em rgba(10, 10, 10, 0.05);
+        max-width: 100%;
+        width: 100%;
+        background-color: white;
+        border-color: #dbdbdb;
+        border-radius: 4px;
+        color: #363636;
+    }
 }
 
 @media (max-width: 1024px) {
