+++ client/src/components/common/Dropdown/index.js
@@ -1,0 +1,4 @@
+import Dropdown from './Dropdown';
+import DropdownItem from './DropdownItem';
+
+export { Dropdown, DropdownItem };
