+++ client/src/styles/base.scss
@@ -1,13 +1,21 @@
 @import './variable.scss';
 @import './progressbar.scss';
 
+$scheme-main: $black;
+$scheme-invert: $white;
+
+html {
+    font-size: 14px !important;
+    scroll-behavior: smooth;
+}
+
 body {
     padding: 0;
     margin: 0;
     min-height: 100vh;
     top: 0;
     overflow: auto;
-}   
+}
 
 .notification {
     z-index: 5;
