+++ client/src/styles/utils.scss
@@ -1,6 +1,9 @@
 .w-100 {
     width: 100%;
 }
+.h-100 {
+    height: 100%;
+}
 
 .flex {
     display: flex;
@@ -9,10 +12,10 @@
 .justify-content-center {
     justify-content: center;
 }
-.justify-space-around {
+.justify-content-around {
     justify-content: space-around;
 }
-.justify-space-between {
+.justify-content-between {
     justify-content: space-between;
 }
